+++ src/lspClient/languageClientMiddleware.ts
@@ -1,10 +1,8 @@
-import { CodeActionMiddleware } from './middlewares/codeActionMiddleware';
 import { Middleware } from './middlewares/middleware';
 
 /**
  * Middleware for {@link LanguageClientCreator}.
  */
 export const LanguageClientMiddleware = {
-  ...CodeActionMiddleware,
   ...Middleware,
 } as const;